requirements.txt
langgraph
sentence-transformers
faiss-cpu
numpy